Создает модуль comments в redux. WIP
